Track cdc stream active time in cdc state table
Summary: Today we maintain and track the stream active time in cache which could result in stale info in the context of leadership changes. So, we plan to write the active time to cdc state table by reading the same from cache at regular intervals as part of UpdateCheckpoint logic. Similarly, whenever a team is found to be inactive by looking into the cache info, we try to confirm the same by reading the active time info captured in the cdc state table.

Test Plan: Test the logic through ctests(part of another diff) which mimics the leadership switches and verifies the expected behaviour.

Reviewers: sdash, skumar, abharadwaj

Subscribers: bogdan

Differential Revision: https://phabricator.dev.yugabyte.com/D19014