[#13693] CDCSDK: Stale entry in CDC Cache causes Steam Expiration.
Summary:
Consider a cluster with 3 tservers (TS1, TS2, TS3) and a table with a single tablet. Today we maintain and track the stream active time in the cache. During starting time tablet LEADER is TS1, so there is a Cache entry for the tablet, to track its active time.  After some time TS2 becomes the tablet LEADER, so an entry will be created in TS2's cache to track the active time of the tablet. now after cdc_intent_retention_ms expiration time,  TS1 becomes a LEADER, but its existing cache entry is not in sync, so if we call GetChanges stream will expire.

To handle this RPC request will be sent by LEADER as part //UpdatePeersAndMetrics// thread, to all the FOLLOWER to update their //last_active_time// in the  CDC Service Cache, so that LEADER and FOLLOWERS are in sync.

Test Plan: Running all the c and java testcases

Reviewers: skumar, srangavajjula, abharadwaj, aagrawal, vkushwaha

Subscribers: ycdcxcluster

Differential Revision: https://phabricator.dev.yugabyte.com/D18882